Changes made by Collin on 11-22-11
Added left and right eye tracking using Haar wavelet
tracking.  There are some new .dll and .xml files that
need to be added to your debug folder in order for the
program to run.  I don't know how to do that other than
to keep a folder of them in the git folder, and copy
them over to the Debug folder after building from the
source file.